+++ telnetlib3/__init__.py
@@ -36,6 +36,8 @@ except ImportError:
     PTY_SUPPORT = False  # invalid-name
 from . import guard_shells as _guard_shells_module
 from .guard_shells import *  # noqa
+from . import sync as _sync_module
+from .sync import *  # noqa
 from .accessories import get_version as __get_version
 # isort: on
 # fmt: on
@@ -54,6 +56,7 @@ __all__ = (
     + telnetlib.__all__
     + (_pty_shell_module.__all__ if PTY_SUPPORT else ())
     + _guard_shells_module.__all__
+    + _sync_module.__all__
 )  # noqa
 
 __author__ = "Jeff Quast"